moved concurrent partitioned tests to own suite
